client/sdk/python/requirements.txt
click>=8.0.0
anthropic>=0.7.0
openai>=1.0.0
google-genai>=0.1.0
